# certif-parent/certif-api-rest/src/main/resources/openapi.yaml
openapi: 3.1.0
...
info:
  title: CertifApp API
  description: |
    API REST de CertifApp — Plateforme SaaS de préparation aux certifications professionnelles.
    
    **Authentification** : JWT Bearer token (access token 15min + refresh token 7j).
    
    **Rôles** : USER (standard), ADMIN (back-office), PARTNER (accès API entreprise).
    
    **Freemium** : Les endpoints marqués `PRO_REQUIRED` retournent 402 si l'abonnement est FREE.
  version: 1.0.0
  contact:
    name: CertifApp Support
    url: https://certifapp.com/support
servers:
  - url: https://api.certifapp.com/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Développement local
security:
  - BearerAuth: [ ]
...
paths:
...
  /auth/register:
    post:
      tags: [ Auth ]
      summary: Créer un compte
      security: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Compte créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /auth/login:
    post:
      tags: [ Auth ]
      summary: Authentification
      security: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/refresh:
    post:
      tags: [ Auth ]
      summary: Rafraîchir les tokens
      security: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ refreshToken ]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Tokens rafraîchis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /certifications:
    get:
      tags: [ Certifications ]
      summary: Liste des certifications disponibles
      security: [ ]
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Liste des certifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CertificationSummaryDto'
  /certifications/{id}:
    get:
      tags: [ Certifications ]
      summary: Détails d'une certification
      security: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: ocp21
      responses:
        '200':
          description: Détails certification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertificationDetailDto'
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Non authentifié
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflit (ex- email déjà existant)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    RegisterRequest:
      type: object
      required: [ email, password ]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          minLength: 8
          example: Str0ngP@ssword
        locale:
          type: string
          default: fr
          example: fr
        timezone:
          type: string
          default: Europe/Paris
    LoginRequest:
      type: object
      required: [ email, password ]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          description: Durée de validité access token en secondes (900 = 15min)
          example: 900
        tokenType:
          type: string
          default: Bearer

    # --- Users ---
...
    CertificationSummaryDto:
      type: object
      properties:
        id:
          type: string
          example: ocp21
        code:
          type: string
          example: 1Z0-830
        name:
          type: string
        totalQuestions:
          type: integer
        passingScore:
          type: integer
        examDurationMin:
          type: integer
        examType:
          type: string
          enum: [ MCQ, PRACTICAL, MIXED ]
    CertificationDetailDto:
      allOf:
        - $ref: '#/components/schemas/CertificationSummaryDto'
        - type: object
          properties:
            description:
              type: string
            themes:
              type: array
              items:
                $ref: '#/components/schemas/CertificationThemeDto'
            examQuestionCount:
              type: integer
    CertificationThemeDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        label:
          type: string
        questionCount:
          type: integer
        weightPercent:
          type: number

    # --- Exams ---
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
        timestamp:
          type: string
          format: date-time
...